fix: avance rediseño de incentivos

+++ servicio-donaciones/src/main/java/com/tp/donatrack/controllers/DonacionController.java
@@ -84,6 +84,6 @@ public class DonacionController {
     @lombok.Getter
     @lombok.Setter
     public static class EntregaRequest {
-        private Integer donacionSegmentadaId;
+        private Long donacionSegmentadaId;
     }
 }
